foodgram/requirements.txt
asgiref==3.4.1
Django==3.2.5
django-debug-toolbar==3.2.1
django-filter==2.4.0
djangorestframework==3.12.4
Pillow==8.3.1
python-dotenv==0.18.0
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
